v0.42
New design: it's gonna be more open world-ish.
Sorry maple, i know that you don't want to hear that I'm doing ANOTHER change this far into my capstone, but I think that this is the level design that'd work best.
I plan on having the boss's office as a sort of level hub, where the player can go to a list of different worlds and then they explore the different towns and complete tasks.

At this point, my game is an amalgamation of different ideas that don't sound like they should go together in any shape or form.

To summarize, my game is likely going to be a world-based completionist game like mario 64 but be a rhythm game on top of it like crypt of the necrodancer / cadence of hyrule. I'm sorry that I didn't think this through more.

I should've just done a fighting game.

Added:
- Reflector enemy
- Imported new swing animation
- Doors for traveling between levels.
- Bus (and bus model) for level select
- Office level, hub level, and first level (Downtown)
- Player now moves on beat, similar to CotND/CoH

Fixes:
- Errors when trying to look at something non-existent. shouldn't throw errors anymore.